requirements.txt
chatterbox-tts==0.1.2
streamlit==1.50.0
gradio
torch==2.6.0
torchaudio==2.6.0
